+++ src/libpocketsphinx/sdc_mgau.h
@@ -84,7 +84,7 @@ void sdc_mgau_free(sdc_mgau_t *s);
 
 int32 sdc_mgau_frame_eval(sdc_mgau_t *s,
                           int16 *senone_scores,
-                          int16 *senone_active,
+                          uint8 *senone_active,
                           int32 n_senone_active,
                           mfcc_t **featbuf,
                           int32 frame,
